+++ workspaces/core/src/group/private/types/GroupComponentProps.ts
@@ -1,0 +1,8 @@
+import type GroupProps from './GroupProps';
+import type WithChildren from './WithChildren';
+
+type GroupComponentProps = (
+  props: WithChildren<GroupProps>
+) => React.ReactElement | null;
+
+export default GroupComponentProps;
