+++ workspaces/core/src/group/private/types/GroupState.ts
@@ -1,0 +1,12 @@
+import type Orientation from './Orientation';
+import type Size from './Size';
+
+type GroupState = {
+  orientation?: Orientation;
+  size?: Size;
+  error?: boolean;
+  disabled?: boolean;
+  readOnly?: boolean;
+};
+
+export default GroupState;
