+++ workspaces/core/src/group/private/types/LabelProps.ts
@@ -1,0 +1,3 @@
+type LabelProps = { className?: string };
+
+export default LabelProps;
