+++ workspaces/core/src/group/private/types/Orientation.ts
@@ -1,0 +1,3 @@
+type Orientation = 'vertical' | 'horizontal';
+
+export default Orientation;
