+++ workspaces/core/src/group/private/types/WithChildren.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react';
+
+type WithChildren<T = {}> = T & { children?: ReactNode };
+
+export default WithChildren;
