+++ workspaces/core/src/group/private/utils/GroupContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+import type GroupState from '../types/GroupState';
+
+const GroupContext = createContext<GroupState>({});
+GroupContext.displayName = 'GroupContext';
+
+export default GroupContext;
